Bug fix : splitting on tokens
An expression such as 5 plus (7 + 2) would not be able to be parsed because it would first try splitting on the + instead of the plus.  The patch fixes this problem by dynamically creating new tokens for each array element (e.g., three new token arrays are created from the plus token array ["\\+", "plus", "add"] : __plus0__ (["\\+"]), __plus1__ (["plus"]) and __plus2__ (["add"]) ) and trying to split on each one separately.